refactor  all nightly tests into two folder (eager and noeager)
