+++ NightlyTests/CMakeLists.txt
@@ -46,8 +46,13 @@ if (ENABLE_TORCH)
 endif (ENABLE_TORCH)
 
 if (ENABLE_TENSORFLOW)
-    add_dependencies(AcceptanceTests
-            AcceptanceTests.Tensorflow)
+    if (NOT TF_VERSION MATCHES "^1.15")
+        add_dependencies(AcceptanceTests
+            AcceptanceTests.TfEager)
+    else()
+       add_dependencies(AcceptanceTests
+            AcceptanceTests.TfNonEager)
+    endif ()
 endif (ENABLE_TENSORFLOW)
 
 if (ENABLE_TORCH)
