+++ NightlyTests/tensorflow/CMakeLists.txt
@@ -57,9 +57,20 @@ add_custom_target(AcceptanceTests.TensorflowDependencies
         "${AIMET_PYTHONPATH}"
         "${Python3_EXECUTABLE}" ${CMAKE_CURRENT_SOURCE_DIR}/dependencies.py ${CMAKE_CURRENT_SOURCE_DIR}/resnet50_eval_scores.csv ${USE_CUDA})
 
-add_custom_target(AcceptanceTests.Tensorflow
+
+
+add_custom_target(AcceptanceTests.TfNonEager
       VERBATIM COMMAND ${CMAKE_COMMAND} -E env
       "${AIMET_PYTHONPATH}"
-      ${Python3_EXECUTABLE} -m pytest -s ${CMAKE_CURRENT_SOURCE_DIR} -m ${MARKER_FLAG} --junitxml=${CMAKE_CURRENT_BINARY_DIR}/py_test_output.xml)
+      ${Python3_EXECUTABLE} -m pytest -s ${CMAKE_CURRENT_SOURCE_DIR}/non_eager -m ${MARKER_FLAG} --junitxml=${CMAKE_CURRENT_BINARY_DIR}/py_test_non_eager_output.xml)
 
-add_dependencies(AcceptanceTests.Tensorflow AcceptanceTests.TensorflowDependencies)
+add_dependencies(AcceptanceTests.TfNonEager AcceptanceTests.TensorflowDependencies)
+if (NOT TF_VERSION MATCHES "^1.15")
+   add_custom_target(AcceptanceTests.TfEager
+      VERBATIM COMMAND ${CMAKE_COMMAND} -E env
+      "${AIMET_PYTHONPATH}"
+      ${Python3_EXECUTABLE} -m pytest -s ${CMAKE_CURRENT_SOURCE_DIR}/eager -m ${MARKER_FLAG} --junitxml=${CMAKE_CURRENT_BINARY_DIR}/py_test_eager_output.xml)
+   add_dependencies(AcceptanceTests.TfEager AcceptanceTests.TensorflowDependencies)
+endif ()
+
+
