+++ NightlyTests/tensorflow/eager/test_bn_reestimation_keras.py
@@ -128,7 +128,6 @@ def _qsim_setup_for_fold_scale(model, dummy_inputs):
 
 
 def _reestimate_and_compare_results(model, dataset):
-
     dummy_inputs = np.random.randn(1, 32, 32, 3).astype(np.float32)
     bn_layers = _get_bn_submodules(model)
     bn_mean_ori = {layer.name: layer.moving_mean.numpy() for layer in bn_layers}
@@ -177,7 +176,6 @@ def test_bn_reestimation():
     batch_size = 4
     dataset = tf.data.Dataset.from_tensor_slices(input_data)
     dataset = dataset.batch(batch_size=batch_size)
-
     dummy_inputs = np.random.randn(2, 32,32,3).astype(np.float32)
     model = tf.keras.applications.mobilenet_v2.MobileNetV2(weights=None, input_shape=(32, 32, 3))
     sub_model = tf.keras.Sequential()
