+++ NightlyTests/tensorflow/eager/test_connected_graph_keras.py
